Cardano.Keys.Hash

Hash

data family Hash keyrole

class CastHash roleA roleB where

Methods

castHash :: Hash roleA -> Hash roleB

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

renderSafeHashAsHex :: SafeHash tag -> Text

HasTypeProxy

class Typeable t => HasTypeProxy t where

Methods

proxyToAsType :: Proxy t -> AsType t

asType :: HasTypeProxy t => AsType t

Provide type proxy from the already existing HasTypeProxy instance

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

data FromSomeType (c :: Type -> Constraint) b where

Constructors

FromSomeType :: forall (c :: Type -> Constraint) a b. c a => AsType a -> (a -> b) -> FromSomeType c b