Cardano.Keys.Serialise.Orphans

Description

Orphan instances the key layer needs and that no upstream package provides.

They belong here, below every module that uses them, so that exactly one copy exists: a second definition elsewhere in the dependency graph would be a duplicate-instance error.

Orphan instances

Data CharPosition

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> CharPosition -> c CharPosition

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c CharPosition

toConstr :: CharPosition -> Constr

dataTypeOf :: CharPosition -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c CharPosition)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c CharPosition)

gmapT :: (forall b. Data b => b -> b) -> CharPosition -> CharPosition

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> CharPosition -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> CharPosition -> r

gmapQ :: (forall d. Data d => d -> u) -> CharPosition -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> CharPosition -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> CharPosition -> m CharPosition

Data DecodingError

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DecodingError -> c DecodingError

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DecodingError

toConstr :: DecodingError -> Constr

dataTypeOf :: DecodingError -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DecodingError)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DecodingError)

gmapT :: (forall b. Data b => b -> b) -> DecodingError -> DecodingError

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DecodingError -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DecodingError -> r

gmapQ :: (forall d. Data d => d -> u) -> DecodingError -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DecodingError -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DecodingError -> m DecodingError

Data DecoderError

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DecoderError -> c DecoderError

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DecoderError

toConstr :: DecoderError -> Constr

dataTypeOf :: DecoderError -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DecoderError)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DecoderError)

gmapT :: (forall b. Data b => b -> b) -> DecoderError -> DecoderError

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DecoderError -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DecoderError -> r

gmapQ :: (forall d. Data d => d -> u) -> DecoderError -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DecoderError -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DecoderError -> m DecoderError

Data DeserialiseFailure

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> DeserialiseFailure -> c DeserialiseFailure

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c DeserialiseFailure

toConstr :: DeserialiseFailure -> Constr

dataTypeOf :: DeserialiseFailure -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c DeserialiseFailure)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c DeserialiseFailure)

gmapT :: (forall b. Data b => b -> b) -> DeserialiseFailure -> DeserialiseFailure

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> DeserialiseFailure -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> DeserialiseFailure -> r

gmapQ :: (forall d. Data d => d -> u) -> DeserialiseFailure -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> DeserialiseFailure -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> DeserialiseFailure -> m DeserialiseFailure

Data UnicodeException

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> UnicodeException -> c UnicodeException

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c UnicodeException

toConstr :: UnicodeException -> Constr

dataTypeOf :: UnicodeException -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c UnicodeException)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c UnicodeException)

gmapT :: (forall b. Data b => b -> b) -> UnicodeException -> UnicodeException

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> UnicodeException -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> UnicodeException -> r

gmapQ :: (forall d. Data d => d -> u) -> UnicodeException -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> UnicodeException -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> UnicodeException -> m UnicodeException

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> UnicodeException -> m UnicodeException

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> UnicodeException -> m UnicodeException

Typeable kd => FromCBOR (VKey kd)

Instance details

Methods

fromCBOR :: Decoder s (VKey kd)

label :: Proxy (VKey kd) -> Text

FromCBOR (OCert StandardCrypto)

Instance details

Methods

fromCBOR :: Decoder s (OCert StandardCrypto)

label :: Proxy (OCert StandardCrypto) -> Text

Typeable kd => ToCBOR (VKey kd)

Instance details

Methods

toCBOR :: VKey kd -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy (VKey kd) -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [VKey kd] -> Size

ToCBOR (OCert StandardCrypto)

Instance details

Methods

toCBOR :: OCert StandardCrypto -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy (OCert StandardCrypto) -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [OCert StandardCrypto] -> Size